Separate out various job classes into their own files
