+++ kge/job/__init__.py
@@ -2,15 +2,15 @@ from kge.job.trace import Trace
 from kge.job.job import Job
 from kge.job.job import TrainingOrEvaluationJob
 from kge.job.train import TrainingJob
-from kge.job.train import TrainingJob1vsAll
-from kge.job.train import TrainingJobKvsAll
-from kge.job.train import TrainingJobNegativeSampling
+from kge.job.train_1vsAll import TrainingJob1vsAll
+from kge.job.train_KvsAll import TrainingJobKvsAll
+from kge.job.train_negative_sampling import TrainingJobNegativeSampling
 from kge.job.eval import EvaluationJob
-from kge.job.eval import TrainingLossEvaluationJob
-from kge.job.grid_search import GridSearchJob
+from kge.job.eval_training_loss import TrainingLossEvaluationJob
+from kge.job.eval_entity_ranking import EntityRankingJob
+from kge.job.eval_entity_pair_ranking import EntityPairRankingJob
 from kge.job.search import SearchJob
-from kge.job.manual_search import ManualSearchJob
-from kge.job.auto_search import AutoSearchJob
-from kge.job.ax_search import AxSearchJob
-from kge.job.entity_ranking import EntityRankingJob
-from kge.job.entity_pair_ranking import EntityPairRankingJob
+from kge.job.search_grid import GridSearchJob
+from kge.job.search_manual import ManualSearchJob
+from kge.job.search_auto import AutoSearchJob
+from kge.job.search_ax import AxSearchJob
